requirements.txt
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
decorator==4.0.10
ipython==4.2.0
ipython-genutils==0.1.0
numpy==1.11.1
pandas==0.18.1
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
pyparsing==2.1.5
python-dateutil==2.5.3
pytz==2016.4
scikits.talkbox==0.2.5
scipy==0.17.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
